+++ internal/sessiontest/legacy_test.go
@@ -1,49 +1,9 @@
 package sessiontest
 
 import (
-	"testing"
-
-	irma "github.com/privacybydesign/irmago"
-	"github.com/privacybydesign/irmago/internal/test"
 	"github.com/privacybydesign/irmago/server/irmaserver"
-	"github.com/stretchr/testify/require"
+	"testing"
 )
-
-func TestSessionUsingLegacyStorage(t *testing.T) {
-	testSessionUsingLegacyStorage(t, "client_legacy")
-}
-
-func TestSessionUsingLegacyPlaintextStorage(t *testing.T) {
-	testSessionUsingLegacyStorage(t, "client_plaintext")
-}
-
-func testSessionUsingLegacyStorage(t *testing.T, dir string) {
-	test.SetTestStorageDir(dir)
-	defer test.SetTestStorageDir("client")
-
-	client, handler := parseStorage(t)
-	defer test.ClearTestStorage(t, client, handler.storage)
-
-	// Test whether credential from legacy storage is still usable
-	idStudentCard := irma.NewAttributeTypeIdentifier("irma-demo.RU.studentCard.studentID")
-	request := getDisclosureRequest(idStudentCard)
-	doSession(t, request, client, nil, nil, nil, nil)
-
-	// Issue new credential
-	doSession(t, getMultipleIssuanceRequest(), client, nil, nil, nil, nil)
-
-	// Test whether credential is still there
-	idRoot := irma.NewAttributeTypeIdentifier("irma-demo.MijnOverheid.fullName.familyname")
-	doSession(t, getDisclosureRequest(idRoot), client, nil, nil, nil, nil)
-
-	// Re-open client
-	require.NoError(t, client.Close())
-	client, _ = parseExistingStorage(t, handler.storage)
-	defer client.Close()
-
-	// Test whether credential is still there after the storage has been reloaded
-	doSession(t, getDisclosureRequest(idRoot), client, nil, nil, nil, nil)
-}
 
 func TestWithoutPairingSupport(t *testing.T) {
 	irmaserver.AcceptInsecureProtocolVersions = true